pyproject.toml
[project]
name = "election69"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "google-genai>=1.59.0",
    "matplotlib>=3.8.0",
    "openai>=2.15.0",
]
